# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * From The Ground Up nonprofit API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Program {
  id: number;
  title: string;
  slug: string;
  description: string;
  longDescription?: string;
  icon: string;
  duration: string;
  skillLevel: string;
  aiFeatures?: string;
  outcomes?: string[];
}

export interface Testimonial {
  id: number;
  name: string;
  program: string;
  quote: string;
  avatar?: string;
}

export type ContactSubmissionType =
  (typeof ContactSubmissionType)[keyof typeof ContactSubmissionType];

export const ContactSubmissionType = {
  enrollment: "enrollment",
  general: "general",
  volunteer: "volunteer",
  donate: "donate",
} as const;

export interface ContactSubmission {
  fullName: string;
  email: string;
  phone?: string;
  programInterest?: string;
  message?: string;
  type: ContactSubmissionType;
}

export interface ContactResponse {
  id: number;
  message: string;
}