# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: From The Ground Up nonprofit API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a single program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Not found
...
components:
  schemas:
...
    Program:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        longDescription:
          type: string
        icon:
          type: string
        duration:
          type: string
        skillLevel:
          type: string
        aiFeatures:
          type: string
        outcomes:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - slug
        - description
        - icon
        - duration
        - skillLevel
...