# lib/api-zod/src/generated/types/program.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * From The Ground Up nonprofit API
 * OpenAPI spec version: 0.1.0
 */

export interface Program {
  id: number;
  title: string;
  slug: string;
  description: string;
  longDescription?: string;
  icon: string;
  duration: string;
  skillLevel: string;
  aiFeatures?: string;
  outcomes?: string[];
}